pyproject.toml
[project]
name = "asr-compare"
version = "0.1.0"
description = "ASR 语音识别引擎对比工具"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "python-multipart>=0.0.6",
    "pydub>=0.25",
    "jiwer>=3.0",
    "dashscope>=1.20",
    "baidu-aip>=4.16",
    "tencentcloud-sdk-python-asr>=3.0",
    "websockets>=12.0",
    "httpx>=0.27",
    "python-dotenv>=1.2.2",
    "chardet>=7.4.3",
    "xfyunsdkspark>=0.0.3",
]
